fix: Position showEmote in display grid, solves element narrowing
The #showEmote element is being reused across consecutive emote showings.
Using the .css() method sets additional properties on the element.

Setting the left, right, top, and bottom properties across subsequent showings restricts the #showEmote element because the opposite properties to the ones being set are not reset.
Setting left to 600 and then on the next showing right to 600 means we effectively restricted the element to left 600 and right 600, with a width of 1280-2*600=80.

Clearing the opposite property would be one solution. But it continues to require an explicit expected rendering width and height, which is problematic / error-prone in-and-of-itself.

Switch to a display grid approach, in this first step with px values, but which can be changed to relative values dissolving the need for a rendering size expectation.

The introduced grid template defines a spacing row and column before and after the img element.
The spacing is adjusted after determining left vs right and top vs bottom alignment. The img area always keeps auto size to support various emote sizes dynamically. The generated x and y values are used as horizontal and vertical spacing size to one of their two sides each (left or right and top or bottom space). The opposite side uses the remaining free space to not influence the positioning.