fix: Transform size multiplier with origin
By default a size transform transforms based on the center of the element.

When a top left 0 0 positioned element is increased in size it will overflow.

By transforming with a transform origin we control into which direction the resizing happens. For a top-left aligned element it will increase towards the bottom right.

This way size increases to not lead to overflows = partially hidden emotes up to an emote size of half width or half height (bigger would overflow if positioned at the center).